refactor(foundation-demo): extract shipping components in checkout page to a shipping-wrapper component
extract shipping components in checkout page to a shipping-wrapper component
