+++ apps/foundation-demo/src/app/checkout/components/shipping-wrapper/shipping-wrapper.component.html
@@ -1,0 +1,13 @@
+<div class="shipping-wrapper" shipping-container #ShippingContainer="ShippingContainer">
+  <shipping-form 
+    [hidden]="!showShippingForm"
+    [shippingInfo]="ShippingContainer.shippingInfo$ | async" 
+    (updateShippingInfo)="ShippingContainer.updateShippingInfo($event); toggleShippingView()"></shipping-form>
+  <shipping-summary 
+    [hidden]="showShippingForm" 
+    *ngIf="ShippingContainer.shippingInfo$ | async" 
+    [shippingInfo]="ShippingContainer.shippingInfo$ | async" 
+    [shippingOption]="ShippingContainer.shippingOption$ | async"
+    (editShippingInfo)="toggleShippingView()"
+    (updateShippingOption)="ShippingContainer.updateShippingOption($event)"></shipping-summary>
+</div>
