+++ apps/foundation-demo/src/app/checkout/components/shipping-wrapper/shipping-wrapper.component.ts
@@ -1,0 +1,19 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'shipping-wrapper',
+  templateUrl: './shipping-wrapper.component.html',
+  styleUrls: ['./shipping-wrapper.component.scss']
+})
+export class ShippingWrapperComponent implements OnInit {
+  
+  showShippingForm: boolean;
+
+  ngOnInit() {
+    this.showShippingForm = true;
+  }
+
+  toggleShippingView() {
+    this.showShippingForm = !this.showShippingForm;
+  }
+}
