+++ apps/foundation-demo/src/app/checkout/pages/checkout-view/checkout-view.component.html
@@ -1,15 +1,3 @@
 <div class="checkout">
-  <div class="checkout__shipping" shipping-container #ShippingContainer="ShippingContainer">
-    <shipping-form 
-      [hidden]="!showShippingForm" 
-      [shippingInfo]="ShippingContainer.shippingInfo$ | async" 
-      (updateShippingInfo)="ShippingContainer.updateShippingInfo($event); toggleShippingView()"></shipping-form>
-    <shipping-summary 
-      [hidden]="showShippingForm" 
-      *ngIf="ShippingContainer.shippingInfo$ | async" 
-      [shippingInfo]="ShippingContainer.shippingInfo$ | async" 
-      [shippingOption]="ShippingContainer.shippingOption$ | async"
-      (editShippingInfo)="toggleShippingView()"
-      (updateShippingOption)="ShippingContainer.updateShippingOption($event)"></shipping-summary>
-  </div>
+  <shipping-wrapper class="checkout__shipping"></shipping-wrapper>
 </div>
