+++ apps/foundation-demo/src/app/checkout/pages/checkout-view/checkout-view.component.ts
@@ -4,17 +4,4 @@ import { Component, OnInit } from '@angular/core';
   templateUrl: './checkout-view.component.html',
   styleUrls: ['./checkout-view.component.scss']
 })
-export class CheckoutViewComponent implements OnInit {
-
-  showShippingForm: boolean;
-
-  constructor() { }
-
-  ngOnInit() {
-    this.showShippingForm = true;
-  }
-
-  toggleShippingView() {
-    this.showShippingForm = !this.showShippingForm;
-  }
-}
+export class CheckoutViewComponent {}
